+++ app/controllers/application_controller.rb
@@ -4,7 +4,7 @@ class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
-  before_action :configure_permitted_parameters, if: :devise_controller?
+  before_action :complete_two_factor_setup
   helper_method :decorated_current_user
 
   rescue_from CanCan::AccessDenied do |exception|
@@ -15,17 +15,40 @@ class ApplicationController < ActionController::Base
   end
 
   def current_user
-    user = super
-    if user
-      User.includes(:roles).find(user.id)
-    end
+    super
+
+    # TODO: this breaks lots of assumptions about current_user
+    # being writeable, and requires more thought.
+    # if user
+    #   User.includes(:roles).find(user.id)
+    # end
   end
 
   def decorated_current_user
     current_user.try(:decorate)
   end
 
+  #TODO: move this into a Warden after_authentication method
+  # This seems like a stop-gap but works technically
+  def complete_two_factor_setup
+    if current_user && !current_user.account_verified && current_user.confirmed_at
+      # Users need to be able to verify their tfa code
+      unless request.path.start_with?(users_two_factor_setup_path)
+        redirect_to new_users_two_factor_setup_path
+      end
+    end
+  end
+
   protected
+  def confirm_two_factor!
+    if current_user.account_verified && !current_user.bypass_tfa
+      # Check identity last checked date
+      if current_user.identity_expired?
+        session[:target_redirect] = request.path
+        redirect_to new_users_two_factor_verification_path
+      end
+    end
+  end
 
   # Calls *stale?* but modifies any supplied *strong_etag* by prepending the value of 
   # ApplicationController#etag_seed.
@@ -37,10 +60,6 @@ class ApplicationController < ActionController::Base
   # ApplicationController#etag_seed.
   def bustable_fresh_when(object = nil, **kwd_args)
     !etag_disabled? && fresh_when(object, strong_etag: bustable_etag(kwd_args[:strong_etag]), **kwd_args)
-  end
-
-  def configure_permitted_parameters
-    devise_parameter_sanitizer.permit(:sign_up, keys: [:first_name, :last_name])
   end
 
   private
@@ -76,4 +95,5 @@ class ApplicationController < ActionController::Base
       editorial_root_path
     end
   end
+
 end