+++ app/controllers/users/authenticator_controller.rb
@@ -1,0 +1,3 @@
+module Users
+
+end