+++ app/validators/aus_phone_number_validator.rb
@@ -1,0 +1,7 @@
+class AusPhoneNumberValidator < ActiveModel::EachValidator
+  def validate_each(record, attribute, value)
+    if value != nil && value !~ /\A04[0-9]{8}\z/
+      record.errors.add(attribute, "is not a valid Australian mobile phone-number")
+    end
+  end
+end