+++ config/environments/development.rb
@@ -59,4 +59,10 @@ Rails.application.configure do
 
   config.version_tag = 'dummy_version'
   config.version_sha = 'dummy_sha'
+
+  # Set SMS provider
+  config.sms_authenticate_url = ENV['SMS_AUTHENTICATE_URL']
+  config.sms_send_message_url = ENV['SMS_SEND_MESSAGE_URL']
+
+  config.force_2fa = ENV['FORCE_2FA']
 end