+++ spec/features/signing_up_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe 'signing up', type: :feature do
 
   let!(:root_node) { Fabricate(:root_node) }
 
-  it 'sets the name' do
+  let(:sign_up) {
     visit new_user_registration_path
     fill_in('Email', with: 'test@example.gov.au')
     fill_in('First name', with: 'first')
@@ -15,12 +15,21 @@ RSpec.describe 'signing up', type: :feature do
     find('#user_password').set('password')
     fill_in('Password confirmation', with: 'password')
     click_button('Sign up')
-
     email_body = Nokogiri::HTML.parse(ActionMailer::Base.deliveries.last.body.raw_source)
     signup_link_url = email_body.css("a")[0][:href]
-
     visit signup_link_url
+  }
 
-    expect(page).to have_content('Your email address has been successfully confirmed')
+
+  describe 'account setup' do
+    context 'user registration' do
+      before {
+        sign_up
+      }
+
+      it 'sets the name' do
+        expect(page).to have_content('Your email address has been successfully confirmed')
+      end
+    end
   end
 end